Deals backend finished
- Deals can be added and delete
- You have a dropdown menu to select which client you wanna associate the deal with
- You can filter clients correctly by many attributes
- Fixed some authorization bugs

+++ app/backend/clientAction.ts
@@ -51,6 +51,12 @@ export async function addClient(state: any, formData: any) {
 }
 
 export async function deleteClient(id: number) {
+  const user = await verifySession();
+  if (!user) {
+    console.error("Unauthorized");
+    return null;
+  }
+  console.log(user);
   try {
     const client = await prisma.client.delete({
       where: { id },
